requirements.txt
ansible==2.9.6
python-libmaas==0.6.6
